Add test_multiple_contribute_inputs() to common
