+++ payjoin/src/core/receive/v1/mod.rs
@@ -331,16 +331,12 @@ mod tests {
 
     use bitcoin::absolute::{LockTime, Time};
     use bitcoin::bip32::{DerivationPath, Fingerprint, Xpriv, Xpub};
-    use bitcoin::hashes::Hash;
     use bitcoin::psbt::Input;
     use bitcoin::secp256k1::Secp256k1;
     use bitcoin::taproot::LeafVersion;
-    use bitcoin::{
-        Address, Amount, Network, OutPoint, PubkeyHash, ScriptBuf, Sequence, TapLeafHash,
-        Transaction,
-    };
+    use bitcoin::{Address, Amount, Network, OutPoint, ScriptBuf, TapLeafHash, Transaction};
     use payjoin_test_utils::{
-        DUMMY20, ORIGINAL_PSBT, PARSED_ORIGINAL_PSBT, QUERY_PARAMS, RECEIVER_INPUT_CONTRIBUTION,
+        ORIGINAL_PSBT, PARSED_ORIGINAL_PSBT, QUERY_PARAMS, RECEIVER_INPUT_CONTRIBUTION,
     };
 
     use super::*;
@@ -741,50 +737,6 @@ mod tests {
             assert!(output.tap_key_origins.is_empty());
             assert!(output.tap_internal_key.is_none());
         }
-    }
-
-    #[test]
-    fn test_multiple_contribute_inputs() {
-        let proposal = unchecked_proposal_from_test_vector();
-        let wants_inputs = wants_outputs_from_test_vector(proposal).commit_outputs();
-        let txout = TxOut {
-            value: Amount::from_sat(123),
-            script_pubkey: ScriptBuf::new_p2pkh(&PubkeyHash::from_byte_array(DUMMY20)),
-        };
-        let tx = Transaction {
-            version: bitcoin::transaction::Version::TWO,
-            lock_time: LockTime::Seconds(Time::MIN),
-            input: vec![],
-            output: vec![txout.clone()],
-        };
-        let ot1 = OutPoint { txid: tx.compute_txid(), vout: 0 };
-        let ot2 = OutPoint { txid: tx.compute_txid(), vout: 1 };
-
-        let input_pair_1 = InputPair::new(
-            TxIn { previous_output: ot1, sequence: Sequence::MAX, ..Default::default() },
-            Input { witness_utxo: Some(txout.clone()), ..Default::default() },
-            None,
-        )
-        .unwrap();
-        let input_pair_2 = InputPair::new(
-            TxIn { previous_output: ot2, sequence: Sequence::MAX, ..Default::default() },
-            Input { witness_utxo: Some(txout), ..Default::default() },
-            None,
-        )
-        .unwrap();
-
-        let wants_inputs = wants_inputs.contribute_inputs(vec![input_pair_1.clone()]).unwrap();
-        assert_eq!(wants_inputs.receiver_inputs.len(), 1);
-        assert_eq!(wants_inputs.receiver_inputs[0], input_pair_1);
-        // Contribute the same input again, and a new input.
-        // TODO: if we ever decide to fix contribute duplicate inputs, we need to update this test.
-        let wants_inputs = wants_inputs
-            .contribute_inputs(vec![input_pair_2.clone(), input_pair_1.clone()])
-            .unwrap();
-        assert_eq!(wants_inputs.receiver_inputs.len(), 3);
-        assert_eq!(wants_inputs.receiver_inputs[0], input_pair_1);
-        assert_eq!(wants_inputs.receiver_inputs[1], input_pair_2);
-        assert_eq!(wants_inputs.receiver_inputs[2], input_pair_1);
     }
 
     #[test]
